Move content from ./src/popup/scripts/ to ./src/popup/
