+++ src/background/actions/action-payloads.ts
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { BaseTelemetryData, TelemetryData, ToggleTelemetryData } from '../../common/telemetry-events';
 import * as TelemetryEvents from '../../common/telemetry-events';
+import { BaseTelemetryData, TelemetryData, ToggleTelemetryData } from '../../common/telemetry-events';
 import { DetailsViewPivotType } from '../../common/types/details-view-pivot-type';
 import { ManualTestStatus } from '../../common/types/manual-test-status';
 import { VisualizationType } from '../../common/types/visualization-type';
 import { TabStopEvent } from '../../injected/tab-stops-listener';
-import { LaunchPanelType } from '../../popup/scripts/components/popup-view';
+import { LaunchPanelType } from '../../popup/components/popup-view';
 
 export interface BaseActionPayload {
     telemetry?: TelemetryData;
