+++ src/background/actions/launch-panel-state-action.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { Action } from '../../common/flux/action';
-import { LaunchPanelType } from '../../popup/scripts/components/popup-view';
+import { LaunchPanelType } from '../../popup/components/popup-view';
 
 export class LaunchPanelStateActions {
     public readonly setLaunchPanelType = new Action<LaunchPanelType>();
