+++ src/background/stores/global/launch-panel-store.ts
@@ -4,7 +4,7 @@ import { autobind } from '@uifabric/utilities';
 
 import { StoreNames } from '../../../common/stores/store-names';
 import { ILaunchPanelStoreData } from '../../../common/types/store-data/ilaunch-panel-store-data';
-import { LaunchPanelType } from '../../../popup/scripts/components/popup-view';
+import { LaunchPanelType } from '../../../popup/components/popup-view';
 import { BrowserAdapter } from '../../browser-adapter';
 import { LocalStorageDataKeys } from '../../local-storage-data-keys';
 import { ILocalStorageData } from '../../storage-data';
