+++ src/common/types/store-data/ilaunch-panel-store-data.d.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { LaunchPanelType } from '../../../popup/scripts/components/popup-view';
+import { LaunchPanelType } from '../../../popup/components/popup-view';
 
 // tslint:disable-next-line:interface-name
 export interface ILaunchPanelStoreData {
