+++ src/popup/components/ad-hoc-tools-panel.tsx
