+++ src/popup/components/diagnostic-view-toggle.tsx
@@ -5,15 +5,15 @@ import { Link } from 'office-ui-fabric-react/lib/Link';
 import { Spinner, SpinnerSize } from 'office-ui-fabric-react/lib/Spinner';
 import { IToggle } from 'office-ui-fabric-react/lib/Toggle';
 import * as React from 'react';
-import { VisualizationToggle } from '../../../common/components/visualization-toggle';
-import { VisualizationConfiguration, VisualizationConfigurationFactory } from '../../../common/configs/visualization-configuration-factory';
-import { KeyCodeConstants } from '../../../common/constants/keycode-constants';
-import { TelemetryEventSource } from '../../../common/telemetry-events';
-import { DetailsViewPivotType } from '../../../common/types/details-view-pivot-type';
-import { IVisualizationStoreData } from '../../../common/types/store-data/ivisualization-store-data';
-import { VisualizationType } from '../../../common/types/visualization-type';
-import { DictionaryStringTo } from '../../../types/common-types';
-import { ContentLink, ContentLinkDeps } from '../../../views/content/content-link';
+import { VisualizationToggle } from '../../common/components/visualization-toggle';
+import { VisualizationConfiguration, VisualizationConfigurationFactory } from '../../common/configs/visualization-configuration-factory';
+import { KeyCodeConstants } from '../../common/constants/keycode-constants';
+import { TelemetryEventSource } from '../../common/telemetry-events';
+import { DetailsViewPivotType } from '../../common/types/details-view-pivot-type';
+import { IVisualizationStoreData } from '../../common/types/store-data/ivisualization-store-data';
+import { VisualizationType } from '../../common/types/visualization-type';
+import { DictionaryStringTo } from '../../types/common-types';
+import { ContentLink, ContentLinkDeps } from '../../views/content/content-link';
 import { PopupActionMessageCreator } from '../actions/popup-action-message-creator';
 import { DiagnosticViewClickHandler } from '../handlers/diagnostic-view-toggle-click-handler';
