+++ src/popup/components/launch-pad-item-row.tsx
