+++ src/popup/handlers/diagnostic-view-toggle-click-handler.ts
@@ -1,13 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as React from 'react';
-
-import { VisualizationConfigurationFactory } from '../../../common/configs/visualization-configuration-factory';
-import { VisualizationActionMessageCreator } from '../../../common/message-creators/visualization-action-message-creator';
-import { TelemetryDataFactory } from '../../../common/telemetry-data-factory';
-import { TelemetryEventSource } from '../../../common/telemetry-events';
-import { IVisualizationStoreData } from '../../../common/types/store-data/ivisualization-store-data';
-import { VisualizationType } from '../../../common/types/visualization-type';
+import { VisualizationConfigurationFactory } from '../../common/configs/visualization-configuration-factory';
+import { VisualizationActionMessageCreator } from '../../common/message-creators/visualization-action-message-creator';
+import { TelemetryDataFactory } from '../../common/telemetry-data-factory';
+import { TelemetryEventSource } from '../../common/telemetry-events';
+import { IVisualizationStoreData } from '../../common/types/store-data/ivisualization-store-data';
+import { VisualizationType } from '../../common/types/visualization-type';
 
 export class DiagnosticViewClickHandler {
     private telemetryFactory: TelemetryDataFactory;
