+++ src/popup/handlers/launch-panel-header-click-handler.ts
