+++ src/popup/handlers/popup-view-controller-handler.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { LaunchPanelType, PopupView } from './../components/popup-view';
+import { LaunchPanelType, PopupView } from '../components/popup-view';
 
 export class PopupViewControllerHandler {
     public openLaunchPad(component: PopupView): void {
