+++ src/popup/launch-pad-row-configuration-factory.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { TelemetryEventSource } from '../../common/telemetry-events';
-import { DetailsViewPivotType } from '../../common/types/details-view-pivot-type';
-import { VisualizationType } from '../../common/types/visualization-type';
+import { TelemetryEventSource } from '../common/telemetry-events';
+import { DetailsViewPivotType } from '../common/types/details-view-pivot-type';
+import { VisualizationType } from '../common/types/visualization-type';
 import { PopupActionMessageCreator } from './actions/popup-action-message-creator';
 import { LaunchPadRowConfiguration } from './components/launch-pad';
 import { PopupView } from './components/popup-view';
