+++ src/popup/main-renderer.tsx
@@ -2,12 +2,11 @@
 // Licensed under the MIT License.
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-
-import { BrowserAdapter } from '../../background/browser-adapter';
-import { Theme, ThemeDeps, ThemeInnerState } from '../../common/components/theme';
-import { WithStoreSubscriptionDeps } from '../../common/components/with-store-subscription';
-import { DropdownClickHandler } from '../../common/dropdown-click-handler';
-import { title } from '../../content/strings/application';
+import { BrowserAdapter } from '../background/browser-adapter';
+import { Theme, ThemeDeps, ThemeInnerState } from '../common/components/theme';
+import { WithStoreSubscriptionDeps } from '../common/components/with-store-subscription';
+import { DropdownClickHandler } from '../common/dropdown-click-handler';
+import { title } from '../content/strings/application';
 import { DiagnosticViewToggleFactory } from './components/diagnostic-view-toggle-factory';
 import { PopupViewControllerDeps, PopupViewWithStoreSubscription } from './components/popup-view';
 import { IPopupHandlers } from './handlers/ipopup-handlers';
