+++ src/popup/popup-initializer.ts
@@ -3,38 +3,36 @@
 import { autobind } from '@uifabric/utilities';
 import { loadTheme } from 'office-ui-fabric-react';
 import * as ReactDOM from 'react-dom';
-
-import { BrowserAdapter } from '../../background/browser-adapter';
-import { A11YSelfValidator } from '../../common/a11y-self-validator';
-import { AxeInfo } from '../../common/axe-info';
-import { VisualizationConfigurationFactory } from '../../common/configs/visualization-configuration-factory';
-import { DropdownClickHandler } from '../../common/dropdown-click-handler';
-import { EnumHelper } from '../../common/enum-helper';
-import { HTMLElementUtils } from '../../common/html-element-utils';
-import { createDefaultLogger } from '../../common/logging/default-logger';
-import { Logger } from '../../common/logging/logger';
-import { ContentActionMessageCreator } from '../../common/message-creators/content-action-message-creator';
-import { DropdownActionMessageCreator } from '../../common/message-creators/dropdown-action-message-creator';
-import { StoreActionMessageCreatorFactory } from '../../common/message-creators/store-action-message-creator-factory';
-import { UserConfigMessageCreator } from '../../common/message-creators/user-config-message-creator';
-import { VisualizationActionMessageCreator } from '../../common/message-creators/visualization-action-message-creator';
-import { AutoChecker } from '../../common/self-validator';
-import { StoreProxy } from '../../common/store-proxy';
-import { BaseClientStoresHub } from '../../common/stores/base-client-stores-hub';
-import { StoreNames } from '../../common/stores/store-names';
-import { TelemetryDataFactory } from '../../common/telemetry-data-factory';
-import { TelemetryEventSource } from '../../common/telemetry-events';
-import { CommandStoreData } from '../../common/types/store-data/command-store-data';
-import { FeatureFlagStoreData } from '../../common/types/store-data/feature-flag-store-data';
-import { ILaunchPanelStoreData } from '../../common/types/store-data/ilaunch-panel-store-data';
-import { IVisualizationStoreData } from '../../common/types/store-data/ivisualization-store-data';
-import { UserConfigurationStoreData } from '../../common/types/store-data/user-configuration-store';
-import { VisualizationType } from '../../common/types/visualization-type';
-import { WindowUtils } from '../../common/window-utils';
-import { contentPages } from '../../content';
-import { ScannerUtils } from '../../injected/scanner-utils';
-import { scan } from '../../scanner/exposed-apis';
-import { SupportLinkHandler } from '../support-link-handler';
+import { BrowserAdapter } from '../background/browser-adapter';
+import { A11YSelfValidator } from '../common/a11y-self-validator';
+import { AxeInfo } from '../common/axe-info';
+import { VisualizationConfigurationFactory } from '../common/configs/visualization-configuration-factory';
+import { DropdownClickHandler } from '../common/dropdown-click-handler';
+import { EnumHelper } from '../common/enum-helper';
+import { HTMLElementUtils } from '../common/html-element-utils';
+import { createDefaultLogger } from '../common/logging/default-logger';
+import { Logger } from '../common/logging/logger';
+import { ContentActionMessageCreator } from '../common/message-creators/content-action-message-creator';
+import { DropdownActionMessageCreator } from '../common/message-creators/dropdown-action-message-creator';
+import { StoreActionMessageCreatorFactory } from '../common/message-creators/store-action-message-creator-factory';
+import { UserConfigMessageCreator } from '../common/message-creators/user-config-message-creator';
+import { VisualizationActionMessageCreator } from '../common/message-creators/visualization-action-message-creator';
+import { AutoChecker } from '../common/self-validator';
+import { StoreProxy } from '../common/store-proxy';
+import { BaseClientStoresHub } from '../common/stores/base-client-stores-hub';
+import { StoreNames } from '../common/stores/store-names';
+import { TelemetryDataFactory } from '../common/telemetry-data-factory';
+import { TelemetryEventSource } from '../common/telemetry-events';
+import { CommandStoreData } from '../common/types/store-data/command-store-data';
+import { FeatureFlagStoreData } from '../common/types/store-data/feature-flag-store-data';
+import { ILaunchPanelStoreData } from '../common/types/store-data/ilaunch-panel-store-data';
+import { IVisualizationStoreData } from '../common/types/store-data/ivisualization-store-data';
+import { UserConfigurationStoreData } from '../common/types/store-data/user-configuration-store';
+import { VisualizationType } from '../common/types/visualization-type';
+import { WindowUtils } from '../common/window-utils';
+import { contentPages } from '../content';
+import { ScannerUtils } from '../injected/scanner-utils';
+import { scan } from '../scanner/exposed-apis';
 import { PopupActionMessageCreator } from './actions/popup-action-message-creator';
 import { DiagnosticViewToggleDeps } from './components/diagnostic-view-toggle';
 import { DiagnosticViewToggleFactory } from './components/diagnostic-view-toggle-factory';
@@ -45,6 +43,7 @@ import { LaunchPanelHeaderClickHandler } from './handlers/launch-panel-header-cl
 import { PopupViewControllerHandler } from './handlers/popup-view-controller-handler';
 import { LaunchPadRowConfigurationFactory } from './launch-pad-row-configuration-factory';
 import { MainRenderer, MainRendererDeps } from './main-renderer';
+import { SupportLinkHandler } from './support-link-handler';
 import { TargetTabFinder, TargetTabInfo } from './target-tab-finder';
 
 declare var window: AutoChecker & Window;
