+++ src/popup/target-tab-finder.ts
@@ -1,11 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { autobind } from '@uifabric/utilities';
-
-import { BrowserAdapter } from '../../background/browser-adapter';
-import { Tab } from '../../common/itab';
-import { UrlParser } from '../../common/url-parser';
-import { UrlValidator } from '../../common/url-validator';
+import { BrowserAdapter } from '../background/browser-adapter';
+import { Tab } from '../common/itab';
+import { UrlParser } from '../common/url-parser';
+import { UrlValidator } from '../common/url-validator';
 
 export interface TargetTabInfo {
     tab: Tab;
