+++ src/tests/unit/tests/background/stores/global/global-store-hub.test.ts
@@ -17,7 +17,7 @@ import { BaseStore } from '../../../../../../common/base-store';
 import { IndexedDBAPI } from '../../../../../../common/indexedDB/indexedDB';
 import { PersistedTabInfo } from '../../../../../../common/types/store-data/iassessment-result-data';
 import { StoreType } from '../../../../../../common/types/store-type';
-import { LaunchPanelType } from '../../../../../../popup/scripts/components/popup-view';
+import { LaunchPanelType } from '../../../../../../popup/components/popup-view';
 import { CreateTestAssessmentProvider } from '../../../../common/test-assessment-provider';
 
 describe('GlobalStoreHubTest', () => {
