+++ src/tests/unit/tests/popup/scripts/components/diagnostic-view-toggle-factory.test.tsx
@@ -13,10 +13,10 @@ import {
 } from '../../../../../../common/configs/visualization-configuration-factory';
 import { TelemetryEventSource } from '../../../../../../common/telemetry-events';
 import { VisualizationType } from '../../../../../../common/types/visualization-type';
-import { PopupActionMessageCreator } from '../../../../../../popup/scripts/actions/popup-action-message-creator';
-import { DiagnosticViewToggle } from '../../../../../../popup/scripts/components/diagnostic-view-toggle';
-import { DiagnosticViewToggleFactory } from '../../../../../../popup/scripts/components/diagnostic-view-toggle-factory';
-import { DiagnosticViewClickHandler } from '../../../../../../popup/scripts/handlers/diagnostic-view-toggle-click-handler';
+import { PopupActionMessageCreator } from '../../../../../../popup/actions/popup-action-message-creator';
+import { DiagnosticViewToggle } from '../../../../../../popup/components/diagnostic-view-toggle';
+import { DiagnosticViewToggleFactory } from '../../../../../../popup/components/diagnostic-view-toggle-factory';
+import { DiagnosticViewClickHandler } from '../../../../../../popup/handlers/diagnostic-view-toggle-click-handler';
 import { ContentLinkDeps } from '../../../../../../views/content/content-link';
 import { ShortcutCommandsTestData } from '../../../../common/sample-test-data';
 import { VisualizationStoreDataBuilder } from '../../../../common/visualization-store-data-builder';
