+++ src/tests/unit/tests/popup/scripts/components/diagnostic-view-toggle.test.tsx
@@ -15,9 +15,9 @@ import { TelemetryEventSource } from '../../../../../../common/telemetry-events'
 import { DetailsViewPivotType } from '../../../../../../common/types/details-view-pivot-type';
 import { IVisualizationStoreData } from '../../../../../../common/types/store-data/ivisualization-store-data';
 import { VisualizationType } from '../../../../../../common/types/visualization-type';
-import { PopupActionMessageCreator } from '../../../../../../popup/scripts/actions/popup-action-message-creator';
-import { DiagnosticViewToggle, DiagnosticViewToggleProps } from '../../../../../../popup/scripts/components/diagnostic-view-toggle';
-import { DiagnosticViewClickHandler } from '../../../../../../popup/scripts/handlers/diagnostic-view-toggle-click-handler';
+import { PopupActionMessageCreator } from '../../../../../../popup/actions/popup-action-message-creator';
+import { DiagnosticViewToggle, DiagnosticViewToggleProps } from '../../../../../../popup/components/diagnostic-view-toggle';
+import { DiagnosticViewClickHandler } from '../../../../../../popup/handlers/diagnostic-view-toggle-click-handler';
 import { DictionaryStringTo } from '../../../../../../types/common-types';
 import { ContentLinkDeps } from '../../../../../../views/content/content-link';
 import { ContentProvider } from '../../../../../../views/content/content-page';
