+++ src/tests/unit/tests/popup/scripts/components/launch-pad-item-row.test.tsx
@@ -7,7 +7,7 @@ import * as React from 'react';
 import * as TestUtils from 'react-dom/test-utils';
 import { Mock, Times } from 'typemoq';
 
-import { LaunchPadItemRow, LaunchPadItemRowProps } from '../../../../../../popup/scripts/components/launch-pad-item-row';
+import { LaunchPadItemRow, LaunchPadItemRowProps } from '../../../../../../popup/components/launch-pad-item-row';
 import { EventStubFactory } from '../../../../common/event-stub-factory';
 
 describe('LaunchPadItemRow', () => {
