+++ src/tests/unit/tests/popup/scripts/components/launch-pad.test.tsx
@@ -5,8 +5,8 @@ import * as React from 'react';
 
 import { ExternalLink } from '../../../../../../common/components/external-link';
 import { toolName } from '../../../../../../content/strings/application';
-import { LaunchPad, LaunchPadDeps, LaunchPadProps, LaunchPadRowConfiguration } from '../../../../../../popup/scripts/components/launch-pad';
-import { LaunchPadItemRow } from '../../../../../../popup/scripts/components/launch-pad-item-row';
+import { LaunchPad, LaunchPadDeps, LaunchPadProps, LaunchPadRowConfiguration } from '../../../../../../popup/components/launch-pad';
+import { LaunchPadItemRow } from '../../../../../../popup/components/launch-pad-item-row';
 
 const AXE_CORE_VERSION = 'axe.core.version';
