+++ src/tests/unit/tests/popup/scripts/components/popup-view-controller-handler.test.tsx
@@ -2,9 +2,9 @@
 // Licensed under the MIT License.
 import { It, Mock, Times } from 'typemoq';
 
-import { PopupActionMessageCreator } from '../../../../../../popup/scripts/actions/popup-action-message-creator';
-import { LaunchPanelType } from '../../../../../../popup/scripts/components/popup-view';
-import { PopupViewControllerHandler } from '../../../../../../popup/scripts/handlers/popup-view-controller-handler';
+import { PopupActionMessageCreator } from '../../../../../../popup/actions/popup-action-message-creator';
+import { LaunchPanelType } from '../../../../../../popup/components/popup-view';
+import { PopupViewControllerHandler } from '../../../../../../popup/handlers/popup-view-controller-handler';
 
 describe('PopupViewControllerHandlerTest', () => {
     test('openAdhocToolsPanel', () => {
