+++ src/tests/unit/tests/popup/scripts/main-renderer.test.tsx
@@ -8,13 +8,13 @@ import { BrowserAdapter, ChromeAdapter } from '../../../../../background/browser
 import { Theme } from '../../../../../common/components/theme';
 import { DropdownClickHandler } from '../../../../../common/dropdown-click-handler';
 import { title } from '../../../../../content/strings/application';
-import { DiagnosticViewToggleFactory } from '../../../../../popup/scripts/components/diagnostic-view-toggle-factory';
-import { PopupViewWithStoreSubscription } from '../../../../../popup/scripts/components/popup-view';
-import { DiagnosticViewClickHandler } from '../../../../../popup/scripts/handlers/diagnostic-view-toggle-click-handler';
-import { LaunchPanelHeaderClickHandler } from '../../../../../popup/scripts/handlers/launch-panel-header-click-handler';
-import { PopupViewControllerHandler } from '../../../../../popup/scripts/handlers/popup-view-controller-handler';
-import { LaunchPadRowConfigurationFactory } from '../../../../../popup/scripts/launch-pad-row-configuration-factory';
-import { MainRenderer, MainRendererDeps } from '../../../../../popup/scripts/main-renderer';
+import { DiagnosticViewToggleFactory } from '../../../../../popup/components/diagnostic-view-toggle-factory';
+import { PopupViewWithStoreSubscription } from '../../../../../popup/components/popup-view';
+import { DiagnosticViewClickHandler } from '../../../../../popup/handlers/diagnostic-view-toggle-click-handler';
+import { LaunchPanelHeaderClickHandler } from '../../../../../popup/handlers/launch-panel-header-click-handler';
+import { PopupViewControllerHandler } from '../../../../../popup/handlers/popup-view-controller-handler';
+import { LaunchPadRowConfigurationFactory } from '../../../../../popup/launch-pad-row-configuration-factory';
+import { MainRenderer, MainRendererDeps } from '../../../../../popup/main-renderer';
 import { SupportLinkHandler } from '../../../../../popup/support-link-handler';
 
 describe('MainRenderer', () => {
